requirements.txt
streamlit
encryption
numpy
rembg
Pillow
